// enums/Pitch.ts
enum Pitch {
  A0 = 'A0',
  A1 = 'A1',
  A2 = 'A2',
  A3 = 'A3',
  A4 = 'A4',
  A5 = 'A5',
  A6 = 'A6',
  A7 = 'A7',
  Ab1 = 'Ab1',
  Ab2 = 'Ab2',
  Ab3 = 'Ab3',
  Ab4 = 'Ab4',
  Ab5 = 'Ab5',
  Ab6 = 'Ab6',
  Ab7 = 'Ab7',
  B0 = 'B0',
  B1 = 'B1',
  B2 = 'B2',
  B3 = 'B3',
  B4 = 'B4',
  B5 = 'B5',
  B6 = 'B6',
  B7 = 'B7',
  Bb0 = 'Bb0',
  Bb1 = 'Bb1',
  Bb2 = 'Bb2',
  Bb3 = 'Bb3',
  Bb4 = 'Bb4',
  Bb5 = 'Bb5',
  Bb6 = 'Bb6',
  Bb7 = 'Bb7',
  C1 = 'C1',
  C2 = 'C2',
  C3 = 'C3',
  C4 = 'C4',
  C5 = 'C5',
  C6 = 'C6',
  C7 = 'C7',
  C8 = 'C8',
  D1 = 'D1',
  D2 = 'D2',
  D3 = 'D3',
  D4 = 'D4',
  D5 = 'D5',
  D6 = 'D6',
  D7 = 'D7',
  Db1 = 'Db1',
  Db2 = 'Db2',
  Db3 = 'Db3',
  Db4 = 'Db4',
  Db5 = 'Db5',
  Db6 = 'Db6',
  Db7 = 'Db7',
  Db8 = 'Db8',
  E1 = 'E1',
  E2 = 'E2',
  E3 = 'E3',
  E4 = 'E4',
  E5 = 'E5',
  E6 = 'E6',
  E7 = 'E7',
  Eb1 = 'Eb1',
  Eb2 = 'Eb2',
  Eb3 = 'Eb3',
  Eb4 = 'Eb4',
  Eb5 = 'Eb5',
  Eb6 = 'Eb6',
  Eb7 = 'Eb7',
  F1 = 'F1',
  F2 = 'F2',
  F3 = 'F3',
  F4 = 'F4',
  F5 = 'F5',
  F6 = 'F6',
  F7 = 'F7',
  G1 = 'G1',
  G2 = 'G2',
  G3 = 'G3',
  G4 = 'G4',
  G5 = 'G5',
  G6 = 'G6',
  G7 = 'G7',
  Gb1 = 'Gb1',
  Gb2 = 'Gb2',
  Gb3 = 'Gb3',
  Gb4 = 'Gb4',
  Gb5 = 'Gb5',
  Gb6 = 'Gb6',
  Gb7 = 'Gb7',
}

export default Pitch